Clase Descontable:     - Corregido el tipo del descuento y desplegados los cambios
Tests:
    - Corregidos los test

+++ src/Descontable.java
@@ -13,13 +13,13 @@ public interface Descontable {
      *
      * @return Valor del atributo descuento
      */
-    int getDescuento();
+    float getDescuento();
 
     /**
      * Método modificador del atributo descuento
      *
      * @param nuevoDescuento Nuevo descuento del precio
      */
-    void setDescuento(int nuevoDescuento);
+    void setDescuento(float nuevoDescuento);
 
 }
